Class refactoring and test failure fixes
