+++ components/src/main/java/com/wso2/openbanking/berlin/extensions/datamodels/ScaApproach.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package com.wso2.openbanking.berlin.extensions.dataobjects;
+package com.wso2.openbanking.berlin.extensions.datamodels;
 
 import com.wso2.openbanking.berlin.extensions.enums.ScaApproachEnum;
