+++ components/src/main/webapp/WEB-INF/beans.xml
